requirements.txt
torch>=2.0.0
transformers>=4.35.0
gradio>=4.0.0
chromadb>=0.4.0
ollama>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
scipy>=1.10.0
sympy>=1.12.0
scikit-learn>=1.2.0
numpy>=1.25.0
Pillow>=10.0.0
EasyOCR>=1.7.0
beautifulsoup4>=4.12.0
cadquery>=2.4.0
datasets>=2.14.0
pylint>=2.17.0
psutil>=5.9.0
tiktoken>=0.5.0
transformers_stream_generator>=0.0.4
einops>=0.8.0
accelerate>=0.26.0
speechrecognition
pyaudio
gtts
pyttsx3
huggingface_hub>=0.29.1
